reorganized analyze.views, added migrate command to docker-compose
