+++ analyze/templates/analyze/fda_smoothing.html
@@ -27,12 +27,12 @@
         </p>
     </div>
 </form>
-<img src="../charts/gcverr">
+<img src="charts/gcverr">
 <h3>
     Optimal log(Lambda) according to minimized error: {{lam}}
 </h3>
 <h3>
     Optimal number of knots according to minimized error: {{knots}}
 </h3>
-<img src="../charts/combined">
+<img src="charts/combined">
 {% endblock %}
