+++ docker-compose.yaml
@@ -20,7 +20,7 @@ services:
           - db
         ports:
           - 8000:8000
-        command: python manage.py runserver 0.0.0.0:8000
+        command: bash -c "python manage.py migrate && python manage.py runserver 0.0.0.0:8000"
 
 volumes:
     postgres_data:
