+++ files/urls.py
@@ -3,7 +3,7 @@ from django.conf.urls.static import static
 from django.urls import path
 from django.conf import settings
 
-from files.views import ProvideFilesView, download_files
+from files.views import ProvideFilesView
 
 urlpatterns = [
     path('', ProvideFilesView.as_view()),
